backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
paho-mqtt
